WRBUF updates

+++ include/yazpp/query.h
@@ -2,12 +2,13 @@
  * Copyright (c) 1998-2005, Index Data.
  * See the file LICENSE for details.
  * 
- * $Id: query.h,v 1.1 2006-03-29 13:14:15 adam Exp $
+ * $Id: query.h,v 1.2 2007-03-20 07:54:11 adam Exp $
  */
 
 #ifndef YAZ_PP_QUERY_H
 #define YAZ_PP_QUERY_H
 
+#include <stddef.h>
 #include <yaz/yconfig.h>
 
 namespace yazpp_1 {
@@ -17,7 +18,7 @@ namespace yazpp_1 {
 class YAZ_EXPORT Yaz_Query {
  public:
     /// Print query in buffer described by str and len
-    virtual void print (char *str, int len) = 0;
+    virtual void print (char *str, size_t len) = 0;
     virtual ~Yaz_Query();
 };
 };
